requirements.txt
# Django and core dependencies
Django>=6.0,<7.0
Pillow>=12.0
requests>=2.33.0
beautifulsoup4>=4.14.0
pypinyin>=0.55.0

# For API calls
httpx>=0.28.0
anyio>=4.0.0
httpcore>=1.0.0
h11>=0.16.0

# For session management (if needed)
# django-redis>=5.0.0  # Uncomment if using Redis for sessions
